requirements.txt
pywebview
cefpython3
qtpy
uvicorn